A variety of small fixes
enums now have static method list() and listr

api models moved under models folder.

Added schema cutting unecessary fields (mostly)

Fixed error handling in API models
