+++ api/app.js
@@ -18,15 +18,11 @@ import cookieParser from "cookie-parser";
 
 const app = express();
 
-app.set("view engine", "ejs");
-
 app.use(cors());
 app.use(helmet());
 app.use(express.json());
-app.use(morgan("tiny"));
 app.use(cookieParser());
 app.use(bodyParser.json());
-app.use("/public", express.static(process.cwd() + "/public"));
 
 app.use("/", pagesRouter);
 app.use("/internal", internalRouter);
