+++ api/error/internal_errors.js
@@ -10,6 +10,12 @@ export class ErrorInternalAPIModelFieldValidation extends InternalError {
   }
 }
 
+export class ErrorInternalAPIModelSchemaValidation extends InternalError {
+  constructor(msg) {
+    super(msg);
+  }
+}
+
 export class ErrorInternalAPIModelValidation extends InternalError {
   constructor(msg) {
     super(msg);
@@ -29,6 +35,24 @@ export class ErrorCannotInitializeAbstractClass extends InternalError {
 }
 
 export class ErrorAbstractMethodNotInstantiated extends InternalError {
+  constructor(msg) {
+    super(msg);
+  }
+}
+
+export class ErrorDatabaseConnection extends InternalError {
+  constructor(msg) {
+    super(msg);
+  }
+}
+
+export class ErrorInternalDatabaseAccessor extends InternalError {
+  constructor(msg) {
+    super(msg);
+  }
+}
+
+export class ErrorInternalUnexpected extends InternalError {
   constructor(msg) {
     super(msg);
   }
