+++ api/models/db_models/photo.js
