+++ api/models/db_models/photo_tag.js
