+++ api/models/enums/article_content.js
@@ -3,12 +3,6 @@ export default class ArticleContent {
   static PullQuote = new ArticleContent("pull_quote");
   static Image = new ArticleContent("image");
 
-  // get all Article Content types
-  static list = [this.BodyParagraph, this.PullQuote, this.Image];
-  static listStr = this.list.map((type) => {
-    return type.type;
-  });
-
   /**
    * INTERNAL USE ONLY
    *
