+++ api/models/enums/design_status.js
@@ -7,12 +7,6 @@ export default class DesignStatus {
   static In_Progress = new DesignStatus("in_progress");
   static Completed = new DesignStatus("completed");
 
-  // Get all DesignStatuses
-  static list = [this.Needs_Designer, this.Has_Designer, this.In_Progress, this.Completed];
-  static listStr = this.list.map((status) => {
-    return status.status;
-  });
-
   /**
    * INTERNAL USE ONLY
    * Construct a DesignStatus enum
