+++ api/models/enums/photography_status.js
@@ -7,12 +7,6 @@ export default class PhotographyStatus {
   static Photographer_Assigned = new PhotographyStatus("photographer_assigned");
   static Photo_Complete = new PhotographyStatus("photo_complete");
 
-  // Get all PhotographyStatuses
-  static list = [this.No_Photo, this.Needs_Photographer, this.Photographer_Assigned, this.Photo_Complete];
-  static listStr = this.list.map((status) => {
-    return status.status;
-  });
-
   /**
    * INTERNAL USE ONLY
    * Construct a PhotographyStatus enum
